Registrar UI displays credits and core now

+++ src/sql/tables/Record.java
@@ -14,7 +14,7 @@ public class Record extends Table {
     public Record() {
         super();
     }
-    public void addRow (int recordId, int average, String honour, String registrationNumber, String periodOfSuty, String registered) {
+    public void addRow (int recordId, double average, String honour, String registrationNumber, String periodOfSuty, String registered) {
         super.addRow(new Object[] {recordId,average,honour,registrationNumber,periodOfSuty,registered});
     }
     public String[] getRegistrationNumbers(){
